[WIP]: Splitting chat component into components
