--- src/components/_copy/Chatbox.scss.txt
@@ -1,12 +1,0 @@
-@import '../../styles/_variables';
-
-.chat-form {
-  position: fixed;
-  bottom: 6rem;
-  left: 4rem;
-  width: calc(100% - 8rem);
-  border-bottom: 2rem solid $main-color-light;
-  height: 8rem;
-  border-radius: $padding-medium;
-  background-color: $main-color-light;
-}
