--- src/components/_copy/IChatTypes.tsx.txt
@@ -1,77 +1,0 @@
-import { QuestionData2 } from './IAxios';
-// export interface ChatEntry {
-//   question: string;
-//   answer: string;
-//   feedback: string;
-// }
-
-// export interface ChatboxProps {
-//   onSubmit: (answer: string) => void;
-//   disabled: boolean;
-// }
-
-// export interface ControlsBoxProps {
-//   onSubmit: (answer: string) => void;
-//   isChatStarted: boolean;
-//   disabled: boolean;
-// }
-
-export interface ReviewQuestion {
-  question: QuestionData2;
-}
-
-// src/types/chat.ts
-
-export interface ChatHistoryItem {
-  question: string;
-  answers: string[];
-  correction: JSX.Element | null;
-  feedback: React.ReactNode | null;
-  selectedAnswer: string | null;
-  questionSetIndex?: number;
-}
-
-export interface ChatUser {
-  name: string;
-  role: string;
-  experience: string;
-  theme?: string;
-}
-
-export interface ChatState {
-  // Chat status
-  isChatStarted: boolean;
-  areControlsDisabled: boolean;
-  areQuestionsLoading: boolean;
-  isAnswerSelected: boolean;
-  isSetCompleted: boolean;
-  showFeedback: boolean;
-
-  // Questions and answers
-  questionSet: QuestionData2[];
-  currentQuestion: QuestionData2 | null;
-  currentQuestionIndex: number;
-  currentAnswers: string[] | null;
-  correctAnswer: string | null;
-  selectedAnswer: string | null;
-
-  // Statistics
-  totalQuestions: number;
-  correctQuestions: number;
-  reviewQuestions: ReviewQuestion[];
-
-  // Chat history
-  chatHistory: ChatHistoryItem[];
-
-  // Actions
-  startChat: (role: string, experience: string, theme: string) => Promise<void>;
-  handleAnswerChange: (answer: string) => void;
-  handleSubmitAnswer: () => void;
-  displayNextQuestion: () => void;
-  startNewQuestionSet: (
-    role: string,
-    experience: string,
-    theme: string
-  ) => Promise<void>;
-  resetChat: () => void;
-}
